+++ ReleaseNotes.md
@@ -2,6 +2,11 @@
 
 # Release Notes CES-Service
 
+## Release 1.1.2
+- dependencies updated
+- try to decode token with url encoding first
+- changed loglevel for token decoding to warn
+
 ## Release 1.1.1
 - No value checks for compositionId for uuid anymore
 - setting new ressources in helm chart
@@ -9,6 +14,7 @@
 - updated dependencies
 - Updated ospo-resources for adding additional notes and disclaimer
 - New realm compatibility
+- Added support for MeinUnternehmenskonto
 
 
 ## Release 1.1.0 (20.02.2025)